etl/requirements.txt
pandas==1.2.3
psycopg2-binary==2.9.2
